% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoard-package.R
\docType{package}
\name{hoardr-package}
\alias{hoardr-package}
\alias{hoardr}
\title{hoardr}
\description{
Manage Cached Files
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
