% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoard-package.R
\docType{package}
\name{hoardr-package}
\alias{hoardr-package}
\alias{hoardr}
\title{hoardr}
\description{
Manage Cached Files
}
\details{
\code{hoardr} is a tiny package with just a single export \code{\link[=hoard]{hoard()}}.
The package goal is to make it easy to setup, use, and manage cached
files for another R package. In addition, you can export functions in
your own package using \code{hoardr} for users to manage their cached
files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/hoardr/}
  \item \url{https://github.com/ropensci/hoardr}
  \item Report bugs at \url{https://github.com/ropensci/hoardr/issues}
}

}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
